+++ src/Extension.php
@@ -11,8 +11,18 @@ trait Extension
     /**
      * @param array<array<int,mixed>|\PHPUnit\Framework\Constraint\Callback<mixed>> $arguments
      */
-    final public static function withConsecutive(array $arguments, mixed $responses = null): Stub
+    protected static function withConsecutive(array $arguments, mixed $responses = null): Stub
     {
         return Assertion::withConsecutive($arguments, $responses);
+    }
+
+    /**
+     * @param list<\PHPUnit\Framework\MockObject\MockObject> $mocks
+     */
+    protected static function neverExpect(array $mocks): void
+    {
+        foreach ($mocks as $mock) {
+            $mock->expects(self::never())->method(self::anything());
+        }
     }
 }